requirements.txt
Flask==2.2.3
openai==0.27.0
python-dotenv==1.0.0
transformers==4.31.0
torch==2.1.0
PyPDF2==3.0.1
python-pptx==0.6.21
python-docx==0.8.11
werkzeug==2.2.3
